refactor: posts utils
